+++ Modules/Utils/Tests/PocketCastsUtilsTests/General/SerialDispatchMock.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-@testable import PocketCastsUtils
-
-final class SerialDispatchMock: DispatchQueueing {
-    func async(execute work: @escaping @convention(block) () -> Void) {
-        work()
-    }
-}