+++ Pocket Casts Watch App Extension/WatchImageHelper.swift
@@ -13,7 +13,7 @@ class WatchImageHelper {
     init() {
         mainCache.diskStorage.config.sizeLimit = UInt(30.megabytes)
         mainCache.diskStorage.config.expiration = .days(45)
-        // the Series 3 Apple watch is slower and has tighter resources, so set our settings there a bit more agressively
+        // the Series 3 Apple watch is slower and has tighter resources, so set our settings there a bit more aggressively
         if DeviceUtil.identifier.lowercased().startsWith(string: "watch3") {
             mainCache.memoryStorage.config.totalCostLimit = Int(1.megabytes)
             mainCache.memoryStorage.config.expiration = .seconds(2.minutes)