+++ podcasts/FeatureFlag.swift
@@ -10,24 +10,6 @@ enum FeatureFlag: String, CaseIterable {
     /// Whether End Of Year feature is enabled
     case endOfYear
 
-    /// Adds the Sign In With Apple options to the login flow
-    case signInWithApple
-
-    /// Displays the new onboarding view updates
-    case onboardingUpdates
-
-    /// Bookmarks / Highlights
-    case bookmarks
-
-    /// Patron
-    case patron
-
-    /// Displaying podcast ratings
-    case showRatings
-
-    /// New episodes autoplay if Up Next is empty
-    case autoplay
-
     /// Enable the new show notes endpoint plus embedded episode artwork
     case newShowNotesEndpoint
 
@@ -48,19 +30,7 @@ enum FeatureFlag: String, CaseIterable {
         case .firebaseLogging:
             return false
         case .endOfYear:
-            return true
-        case .signInWithApple:
-            return true
-        case .onboardingUpdates:
-            return true
-        case .bookmarks:
-            return true
-        case .patron:
-            return true
-        case .showRatings:
-            return true
-        case .autoplay:
-            return true
+            return false
         case .newShowNotesEndpoint:
             return false
         case .episodeFeedArtwork: