+++ podcasts/NowPlayingPlayerItemViewController.swift
@@ -44,6 +44,11 @@ class NowPlayingPlayerItemViewController: PlayerItemViewController {
     @IBOutlet var podcastName: ThemeableLabel! {
         didSet {
             podcastName.style = .playerContrast02
+            let tapGesture = UITapGestureRecognizer(target: self, action: #selector(podcastNameTapped))
+            podcastName.addGestureRecognizer(tapGesture)
+
+            podcastName.accessibilityTraits = .button
+            podcastName.accessibilityHint = L10n.accessibilityHintPlayerNavigateToPodcastLabel
         }
     }
 
@@ -118,6 +123,7 @@ class NowPlayingPlayerItemViewController: PlayerItemViewController {
 
     @IBOutlet var timeSlider: TimeSlider! {
         didSet {
+            timeSlider.accessibilityLabel = L10n.accessibilityEpisodePlayback
             timeSlider.delegate = self
         }
     }
@@ -176,7 +182,7 @@ class NowPlayingPlayerItemViewController: PlayerItemViewController {
         super.viewDidAppear(animated)
 
         // Show the overflow menu
-        if FeatureFlag.bookmarks.enabled, AnnouncementFlow.current == .bookmarksPlayer {
+        if AnnouncementFlow.current == .bookmarksPlayer {
             overflowTapped()
         }
     }
@@ -285,6 +291,11 @@ class NowPlayingPlayerItemViewController: PlayerItemViewController {
 
     @objc private func chapterNameTapped() {
         containerDelegate?.scrollToCurrentChapter()
+    }
+
+    @objc private func podcastNameTapped() {
+        Analytics.track(.playerPodcastNameTapped)
+        containerDelegate?.navigateToPodcast()
     }
 
     private func skipForwardLongPressed() {