+++ podcasts/Strings+Generated.swift
@@ -40,8 +40,12 @@ internal enum L10n {
   internal static var accessibilityDisabled: String { return L10n.tr("Localizable", "accessibility_disabled") }
   /// Dismiss
   internal static var accessibilityDismiss: String { return L10n.tr("Localizable", "accessibility_dismiss") }
+  /// Episode Playback
+  internal static var accessibilityEpisodePlayback: String { return L10n.tr("Localizable", "accessibility_episode_playback") }
   /// Tap to hide filter details
   internal static var accessibilityHideFilterDetails: String { return L10n.tr("Localizable", "accessibility_hide_filter_details") }
+  /// Tap to navigate to main podcast information page
+  internal static var accessibilityHintPlayerNavigateToPodcastLabel: String { return L10n.tr("Localizable", "accessibility_hint_player_navigate_to_podcast_label") }
   /// Double tap to star episode
   internal static var accessibilityHintStar: String { return L10n.tr("Localizable", "accessibility_hint_star") }
   /// Double tap to remove star from episode
@@ -53,6 +57,10 @@ internal enum L10n {
   /// %1$@ percent completed
   internal static func accessibilityPercentCompleteFormat(_ p1: Any) -> String {
     return L10n.tr("Localizable", "accessibility_percent_complete_format", String(describing: p1))
+  }
+  /// %1$@ of %2$@
+  internal static func accessibilityPlaybackProgress(_ p1: Any, _ p2: Any) -> String {
+    return L10n.tr("Localizable", "accessibility_playback_progress", String(describing: p1), String(describing: p2))
   }
   /// Playback speed %1$@ times
   internal static func accessibilityPlayerEffectsPlaybackSpeed(_ p1: Any) -> String {